build_zynqmp_boot_bin.sh: Update for DDR4 SODIMM change.
This change is targeting our supported projects based on zcu102,
more precisely, this is for the users owning a zcu102 rev1.1
starting from label 0432055-05.

" - Due to an End of Life notification from Micron Technology, Inc.
the DDR4 SODIMM part on the ZCU102 Evaluation Kit and the ZCU106
Evaluation Kit has changed.
 - The change in the component width means that different memory
settings are required during the FSBL for designs to function.
- Starting in Vivado 2018.3, when targeting a ZCU102/ZCU106 board,
the FSBL will have a custom function to query the SPD prom on the
DIMM to determine which DIMM is being used." This is from AR# 71961.

More info on:
- https://www.xilinx.com/support/answers/71961.html
- https://www.xilinx.com/support/answers/72113.html

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>